Bugfix: Parsed too little numbers from the power_supply file
The values in the /sys/class/power_supply/BAT0/uevent are 8 numbers long.
But the class only read 5 of them.
So if the enery decreased from e.g.: 10000(000) to 90000(00) the value suddely became bigger than it really was. That shows in the -s report several error and it tooks me a long time to find this bug.